+++ CHANGES.md
@@ -2,6 +2,13 @@
 This files lists the changes during the lifetime of this project.
 
 ## unreleased
+- add period filter to placement page
+- add end date of current placement to colleague list
+- add availability sorting to colleague list
+- (backwards incompatible) change assignment status. new list: LEAD, VACATURE, INGEVULD, AFGEWEZEN
+- changed that assignment phase is computed instead of assigned
+- changed to also unavailble colleagues are shown when checking out matches
+- fix source data to have correct Placement.period_source
 
 ## 2025-09-19
 - update dependencies to latest, including django security release